p2p: Relocate Server symbols across dedicated files
